Bug 1277955 - Add support for revision_hash to pulse jobs
Some repos are longer-lived and do not yet have the Task Cluster
code that allows them to submit tasks with a revision.  They only
have the older code to submit revision_hash.  This prevents the
jobs from being ingested via Pulse.  This commit adds support
for revision_hash until a time when it's no longer needed.