Bug 1277575 - Don't fail the task when autoclassification encounters a duplicate FailureMatch
Refactor the code slightly by forward-porting the changes from the
elasticsearch branch that cause us to optimise out further matching when
we find a match that is considered good enough. Also wrap the creation
of a FailureMatch in a try/except block so that we don't fail if there
is an existing FailureMatch for the same line/classification which could
happen if the task previously failed for an unrelated reason. This gives
higher granularity than just putting the entire task in a transaction.