Bug 1278962 -  Match crash signatures first by same test then irrespective of test
Instead of matching all crash signatures and then sorting by whether the
test is equal to the test for the new failure line, first try to match
with the same test and then try to match with any test iff there was no
result from the first query.

This change adds an index to make (signature, test) matches more efficient.