Bug 1260805 - Make performance bug template configurable per-project
Do this by storing the content of each template in the database (currently
we have just one, for Talos) and adding an API endpoint for fetching it.
For now, we will store the data that should be in each template in a
pre-loaded fixture. Bonus new feature: automatic population of a "CC" list.