Bug 1116372 - Stop presenting "intermittent needs filing" as a classification option type
This seemed like a good idea in theory, but in practice is not very useful
without adding a new view that only shows failures classified as this
type, so we might as well drop the type.

This patch only prevents new failures from seeing it as an available
classification type. It does not migrate any existing failures classified
as this type or remove the type from the model.