Bug 1279213 - Python client: Remove the per-method timeout parameters
The `timeout` parameter to requests.{get,post} only control the timeout
waiting for a response to the socket, and not the total request time. As
such, it doesn't really make sense to want to specify different timeouts
for different API requests (since the time to connect should be similar,
even if one API request is more time-consuming to generate overall).

This commit removes the per-method timeout argument, leaving the
TreeherderClient constructor as the single place to control the timeout.