Bug 1277304 - Migration-day changes for SCL3 only [DO NOT MERGE]
On migration day, this branch will be deployed to SCL3 (but not to
Heroku) to prepare for the switchover to Heroku. These changes will
never be merged to master.

During the migration, the celery queues must reach zero, since unlike
the MySQL database, anything left in the SCL3 rabbitmq instance will not
be transferred to the Heroku one.

At the point at which the DB replication is shut off, we must also
ensure no new DB writes are occurring in SCL3.

As such, this commit:
* Adds a maintenance page served instead of all UI assets (served with
an HTTP 200 status code, rather than 503, so Zeus doesn't take the
webheads out of rotation), so users are informed of the downtime.
* Disables the celery beat process, so new periodic tasks (such as
pushlog/builds-4hr fetching, cycle_data etc) are not scheduled.
* Disables the pulse ingestion process (we're only using this on stage
currently, but may be on prod soon), to stop pulling in new Taskcluster
jobs from their Pulse queue. Jobs will safely build up in Pulse in the
meantime.
* Disables API changes made by any third party Hawk user (eg taskcluster
or autophone), to prevent jobs table writes and resultant log parser
tasks.
* Disables API changes made by Django-authenticated users (eg for the
sheriffing visibility profiles or classifying failures).
* Allows the `treeherder-etl` Hawk client id to continue to make changes
via the API, so that the Treeherder ETL tasks (eg log parsing) can
continue to run, to empty the queues out. (The queues will soon empty,
due to the celery beat task and third party API changes being disabled)