Bug 1283865 - remove pattern requirements for pulse job display fields (#1639)
Several existing jobs are already out of compliance with these patterns
and there is no existing way to tell task definition developers how to
comply with our required patterns.

created bug 1283866 in Taskcluster for that tool/workflow

This removes the pattern requirements.  If we ever decide that we DO
need these patterns, we can create that tool and then fix old task
definitions to comply.