Bug 1284990 - Use user email address instead of username (#1654) r=emorley
I originally used usernames so I wouldn't be printing full email addresses
into bugs. Apparently older users in Treeherder have a readable name as
the username field, but newer users have a random string of
letters/numbers, which is much less useful.

This change makes it so we print the email address in the form 'user [at]
domain.com', which is slightly better.