Bug 1284911 - Adding check to avoid confusion with old action tasks.
Action Tasks are tasks which will be generated when the "Add New Jobs" button is pressed, and TaskCluster jobs are scheduled.
Action Tasks will contain all the relevant logs. An orange Action Task would mean "Add New Jobs" has failed for Action Tasks.

Action Tasks are placed under the gecko-decision group. This check improves the filter which earlier assumed that there was just a single job under gecko-decision.
Hence when we've scheduled an Action Task on that push in the past, this filter will be necessary to avoid internal server errors.