WIP Bug 1123814 - Desktop notifications
When the new actionmenu button is pressed, the selected revision is added
to a list and an interval function is started.

When the interval function is run, each revision in the list is checked
for percentComplete. If that is 100%, a desktop notification is shown and
the revision is removed from the list.

This only works while the page is currently loaded. A future PR could
implement a serviceworker that could run outside of the page, but that's
more than I want to work on for a WIP.