Bug 1288181 - Handle empty changeset list from json-pushes (#1717)
If a pushlog contains hidden changesets (changesets that are
obsolete) then the call to json-pushes will return a push
with no changesets.  This was changed in bug 1286426.
For us, if `changesets` is empty, we will not be able to get
a revision, which is required for a resultset.  So we
need to just skip the push.