Bug 1275405 - test_auth.py: Test the `...OrReadOnly` permissions_class
The tests previously used the `HasHawkPermissions` permissions class,
which is not actually used in Treeherder at all. More useful would be to
test `HasHawkPermissionsOrReadOnly`, since that is used. The only
difference is that for the latter, read-only requests (such as GETs) are
allowed to succeed even if no credentials were provided.

Note however, that even if a request is read-only, if incorrect
credentials are given (or that user isn't 'authorised'), then the
request won't succeed regardless.

Since anonymous GETs are now allowed to succeed, the expected response
content has been changed to something more generic to avoid confusion.

This change makes `HasHawkPermissions` unused, and so it will be merged
into `HasHawkPermissionsOrReadOnly` later.