Bug 1275405 - Clarify error message when Hawk client_id not found
The previous message implied the client_id wasn't known at all, when in
reality it could either be unknown or just not authorised. In the future
we may allow authenticating even if the credentials are not 'authorised'
(for example for higher rate limits for GETs, even if the user isn't
permitted to make POSTs to job endpoints etc), but for now we should try
and avoid confusion.