Bug 1277304 - Migration-day changes for SCL3 only [DO NOT MERGE]
On migration day, this branch will be deployed to SCL3 (but not to
Heroku) to prepare for the switchover to Heroku. These changes will
never be merged to master.

During the migration, the celery queues must reach zero, since unlike
the MySQL database, anything left in the SCL3 rabbitmq instance will not
be transferred to the Heroku one.

At the point at which the DB replication is shut off, we must also
ensure no new DB writes are occurring in SCL3.

As such, this commit:
* Adds a dismissible maintenance overlay to the main Treeherder page,
so users are informed of the downtime. The login buttons are also
disabled, and made to re-show the maintenance info.
* Blocks any requests to the site that are not of type GET/HEAD/OPTIONS,
from any user (whether Django or Hawk authenticated). This will prevent
job submissions as well as user failure classifications/admin panel
changes and so on.
* Disables the celery beat process, so new periodic tasks (such as
pushlog/builds-4hr fetching, cycle_data etc) are not scheduled.
* Disables the pulse ingestion process, to stop pulling in new
Taskcluster jobs from their Pulse queue. Jobs will safely build up in
Pulse in the meantime.