Bug 1291010 - Add branch field to Repository model/table (#1755)
We need a 'branch' field on the repository so that we can determine
which repo to use for incoming resultsets from pulse exchanges. In
the past, projects like gaia-taskcluster have had their own maps of
github repositories/names/branches to Treeherder projects.  But
Treeherder should be the one owning that mapping.  The only thing on
this table was the branch that's used.  So here it is.

The "branch" field will default to "master" which is appropriate for
several of the repos.  But a few will need more custom values set.
These are laid out in the fixtures/repository.json file.  But they
will need to be manually entered into the databases on Prod/Stage
and Heroku.