Bug 1298539 - Make sure canceling a single job only works from the same push
When you select a single job and then press the cancel all jobs button,
the behavior of the cancel all jobs button so it cancels just the selected
job.

Prior to this patch, the selected job can be canceled no matter which
cancel all jobs button is clicked, even if the selected job is from a
different push. This push only activates the selected job behavior when
the selected job is from the push where the cancel all jobs button was
pressed.