Bug 1290082 - Make ThModelErrors use HTTP status code if no API message
Previously if the server response wasn't valid json with a `detail`
attribute, only the generic message from the `ThModelErrors.format()`
call (eg "Unable to cancel job") would be displayed.

Now, `format()` will fall back to including the HTTP status code (and
corresponding text description of that code) too.

We intentionally don't display the server response if it's not valid
json, since in this case it's likely several hundreds lines of HTML from
the load balancer. (And users can always inspect the response manually.)