Bug 1286578 - Retry job task if resultset doesn't exist (#1872)
* Bug 1286578 - Retry job task if resultset doesn't exist

This removes the logic which creates `skeleton resultsets`
when a job is ingested that we don't have a resultset for yet.
The new approach is to fail and wait for the task to retry.

The buildbot job ingestion already skips and retries later if
it encounters a job for which it has no resultset.

This adds a similar check to the Pulse Job ingestion.  If
a job comes in with a revision that doesn't have a resultset
yet, then this will raise a ValueError.  That will invoke the
retryable_task actions which will wait a bit, then retry.  Each
time it will wait a little longer to retry.  After 9 retries it
waits something like 3900 seconds which should be plenty of time
for the resultset ingestion to complete.