Bug 1306889 - Help ensure bugfiler doesn't choke on xperf failures
For whatever reasons, xperf failures are logged as "TEST-UNEXPECTED-FAIL :
xperf :" instead of "TEST-UNEXPECTED-FAIL | xperf |", which caused issues
with the bugfiler trying to split the failure line on |s. This patches
over the problem by not letting the summary array be reduced if there's
only one item left in the array.