Bug 1308669 - Make sure web-platform-tests is suggested for wpt jobs
This just makes sure that when the selected failing job is a wpt test, the
Testing :: web-platform-tests product/component pair is offerred as a
suggestion.

This also helps make sure we don't try to use "TEST-UNEXPECTED-foo" as a
potential path root for the initial product searches, because that makes
the feature fail to find anything if the "TEST-UNEXPECTED-foo" isn't
omitted.