Bug 1220164 - Add a simple collapse/expand button per push
This adds a button to the left of each push header, allowing the user to
collapse or expand all details about this push. It does this by setting an
attribute on the push as a whole, and hiding non-header parts of the push
when that attribute is set. Clicking the button again will remove that
attribute from the push, causing the rest of the push's information to be
displayed again.