Bug 1310974 - Perform autoclassification after crossreferencing error lines (#1955)
In the short term this change is a small negative as it will slow down
autoclassification tasks. However it's a necessary step to future
changes that will put the autoclassification results on the text log
error table and allow autoclassification of non-structured lines.

As part of this, we remove the use of callback tasks and just schedule
the crossreference task to run after the parser tasks. This is simpler
and reflects the fact that we only schedule one parser task at a time
now.