Bug 1317682 - Use the domain from SITE_URL to set ALLOWED_HOSTS
One of the Django 1.8.16 security changes means that `ALLOWED_HOSTS` is
now used even when `DEBUG` mode is enabled. As such, we must now set it
in the Vagrant environment.

However we don't really need to be able to set it independently of
`SITE_URL`, so we might as well derive it from that.