Bug 1273034 - Transition to Taskcluster Auth from Persona (#1922)
In this commit, Sheriff access is still maintained in the
Treeherder DB, rather than using the scopes derived from
LDAP.

For local usage with Vagrant, this requires accessing
Treeherder with localhost instead of
local.treeherder.mozilla.org

Loggin in to the Django Admin is not enabled in this
branch.  Do use the admin, you must first login through
the normal Treeherder front-end.  Then the admin will
be accessible if the user has the privileges to do so.

Persona login will still be technically possible through the
login.taskcluster.net site.  But that choice will go away
shortly.