Bug 1328618 - Make SETA adjust job priorities in both directions
Up until now, job priorities could only be increased to high value jobs.
This change allows resetting a job to a low value priority if it has not
been determined to be a high value job.

This change also fixes a race condition between load_preseed and
analyze_failures where we would attempt to update the job priority of a
job that is not yet in the database.