Bug 1320136 - Mark redundant RunSQL migration operations as elidable
Django's `squashmigrations` command cannot optimise past `RunSQL` or
`RunPython` migration operations, since it cannot guarantee that the
order of migration operations isn't important. However as of Django 1.10
these commands can be marked as `elidable`, if they are not required in
the final migration (eg if they only manipulate legacy data and have
already run in all necessary environments), which solves this problem.