Bug 1320136 - Squash Django migrations
This reduces the time for Vagrant provision & also stops us hitting
some Django migration bugs around index changes (which was causing
redundant indexes on fields that were unique keys as well as a missing
index on `signature_hash`).

Django has a built-in migrations squashing feature:
https://docs.djangoproject.com/en/1.10/topics/migrations/#squashing-migrations
...however it cannot optimise across RunSQL commands and several other
cases, so results in poorly optimised migrations. As such these squashed
migrations are a combination of the output of squashmigrations:

./manage.py squashmigrations credentials 0003
./manage.py squashmigrations model 0053
./manage.py squashmigrations perf 0032
./manage.py squashmigrations seta 0002

...supplemented with the output from deleting all migrations and
running `./manage.py makemigrations` from scratch.

For apps that have already run all migrations that this squashed
migration replaces, this migrations file is a no-op.

Once enough time has passed for all environments (including local
Vagrant instances) to be up to date with the squashed migration, the
old migrations files can be deleted.