Bug 1245389 - Log viewer: Display log lines prior to the selected line
Currently if you click on a failure, the viewer would put the red
failure line on the very top of the log area, which is less optimal
since some log above the failure line could usually be useful. This
commit will add an offset of -7px to a line number so that users can
see the log above a failure line.