Bug 1333422 - Remove useless resize-watching code from ui
We have some code which tries to get the size of a non-existent "top navbar"
to adjust the size of the resultset/jobs view. I think this dates way back
to 2014, before I added flexbox-- the code is harmless (since it does
nothing), but causes a confusing console error ("Empty string passed to
getElementById"). Let's remove it.