Bug 1294354 - Modify log style to use line class (#2109)
Copy-pasting from the logviewer introduces empty lines in between. This
is because we are using `<p>` tags to wrap a line in the viewer.
Switching `<p>` to `<div>` fixes the problem. Since the logviewer within
Treeherder is styling based on tags, we will need to do this change in 3
steps in order to avoid downtime. This PR does the **second** step.

Steps:
1. **logviewer-pr**: Change `<p>` to `<p class="line...>`
2. **th-pr**: Change Treeherder logviewer styling to use `.line` instead of `p`
3. **logviewer-pr**: Change `<p class="line...>` to `<div class="line...>`