Bug 1304062 - Remove prefix index on failure_line.signature
Since it overlaps with the composite index, and the additional length
isn't required (50 characters vs 25).

The RunSQL operation is being removed from the existing migrations file
(rather than adding a new RunSQL operation that reverts it), since the
index is missing on prod, so any attempts to remove it would error. This
means that existing Vagrant environments will have an extra index until
the environment is recreated, however it's unlike to have an impact on
reproducibility of prod issues.