Bug 1284432 - Cycle jobs after failure line data
For historical reasons, the failure line data is currently related to the job
via the guid (not a foreign key relation). To avoid the case where we have
orphan data where we die halfway through deleting the job data, make sure
we cycle the failure line data *before* any jobs.