Bug 1347946 - Add pinboard integration to the Failure Classification panel
Currently non-sheriffs cannot classify through the Failure
Classification panel. It turns out that there are a few peoeple who
want to do that. The simplest approach is to add the pinboard icons
analogous to the old UI so that they can mark jobs as classified but
not affect the autoclassification data. For sheriffs the pinboard is
also important for non-intermittents, so add a bustage button to the
toolbar and show the pinboard icons when the job is pinned.