Bug 1353014 - Move contribute.json to ui/ to simplify build config
contribute.json is intended for consumption via the deployed site, so
does not need to be in the repo root. Everything in `ui/` will be copied
to `dist/` so once moved, `contribute.json` no longer needs to be listed
explicitly - and with Neutrino v5 this will allow for even more
simplification later.