Bug 1345131 - Update pytest-* packages
Since the new pytest-html version now depends on pytest-metadata, and
pyup doesn't yet handle adding new sub-dependencies as part of package
updates, so we have to do so manually. The other packages have to be
updated at the same time, since pytest-html made breaking changes in a
minor version, sigh.

https://github.com/pytest-dev/pytest-html/blob/master/CHANGES.rst
https://github.com/pytest-dev/pytest-html/compare/v1.13.0...v1.14.2

https://github.com/pytest-dev/pytest-variables/blob/master/CHANGES.rst
https://github.com/pytest-dev/pytest-variables/compare/v1.4...v1.5.1

https://github.com/pytest-dev/pytest-base-url/blob/master/CHANGES.rst
https://github.com/pytest-dev/pytest-base-url/compare/v1.2.0...v1.3.0

https://pytest-selenium.readthedocs.io/en/latest/news.html
https://github.com/pytest-dev/pytest-selenium/compare/v1.8.0...v1.9.1