Bug 1353093 - Make migration 0008 a no-op (#2334)
Django bug https://code.djangoproject.com/ticket/28065 causes the
addition of ``related_name`` to create a migration that will drop and
re-add the foreign key.  However, this is a ``meta`` only change, so
no DB changes should be made and this ForeignKey drop and re-add
is super expensive.  I didn't notice this problem before I merged
commit d85b8f74d2bd63a7bd1de420aa15f422b360dbc8.