Bug 1356047 - Allow csrf_exempt on graphql endpoints (#2340)
This is necessary to allow the TestGroup UI to access
these GraphQL endpoints from a separate domain.  The
client library we are using is Apollo, which uses a POST,
even for read-only requests.  It is easier to ease
the CSRF protections here than to get Apollo to support
GET requests at this time.  We'll likely move to GET
requests at a later time.

Disabling CSRF protections on these endpoints is safe
because they are all read-only.  If we create "mutations"
for these endpoints which allow writes, we'll need to
re-visit the use of this exemption.