Bug 1366345 - Force the save() event to be emitted when ctrl-enter is used while manually entering bug numbers (#2494) r=camd
Something is eating the event fired when ctrl-enter is pressed to save classifications if the manual bug number field is focused, making the user press ctrl-enter twice to save the classification.  This patch adds a keypress listener to the bug entry field which forces the save() function to get called when ctrl-enter is pressed.