Bug 1355339 - Make the bug filer warn when it doesn't match any of the failure's search terms (#2336) r=emorley
If the selected failure has no search_terms defined, the bug being filed will never appear as a suggestion for classification. Warn in this case.

If the bug-to-be-filed's summary does not contain any of the selected failure's search_terms, it will never appear as a suggestion for the classification. Warn in this case. The search terms are displayed in this case so the user can see exactly what needs to be added to make the bug discoverable.