Update isort from 4.2.5 to 4.2.14
The new version treats whitespace slightly differently, and also
counts `concurrent.futures` as a stdlib (since it is in Python 3,
even though we have to use the `futures` package). However I'm fine
with not overriding the latter for simplicity and given we'll be
switching to Python 3 at some point in the future.