Bug 1375712 - Return bug suggestions from graphql queries
This GraphQL Schema addition is a little different in that we are
enabling the return of an object that is not represented by a Django
Model.  The Bug Suggestions is fetched from the cache and returned
as an object.  So I've introduced a new ``ObjectScalar`` to represent
a simple ``Object`` type to be returned.