Convert log parsing into a single task.
Convert the log parsing pipeline into a single task that parses
unstructured logs, stores error lines and crossreferences the
structured and unstructured lines. This relies on the fact that we
always get all the error lines for a job at once, and therefore don't
need to consider the case where a job gets structured and unstructured
log data in different requests.

The patch has the effect of unbreaking the crossreference_error_lines
functionality which means that autoclassification might start to do
something more useful again. It also has the downside that the
log_parser task now does more work in serial so the latency for a
specific job may be increased.