Bug 1335920 - support both $dumps and $json
JSON-e defines `{$json: ..}`, and that's what the actions documentation
says, but Treeherder has been expecting `$dumps`, and the current
`action.json` files use `$dumps`.  So this is the first step toward
migrating to the JSON-e standard: make TH accept both.  Once this is
deployed, we can switch the in-tree code to generate `$json`.  A little
while later, we can switch TH to use JSON-e instead.