Bug 1372581 - Better new relic reporting for resultset_loader
Add fields of ``url`` and ``repo`` in case exceptions happen.
This was catching an exception that could happen in
``fetch_resultset`` for github pushes, but wasn't raising it, so
we would get exceptions later due to a null object returned
from that function.  This allows the exception to raise and
not try to do any more processing on that push.