Bug 1378473 - Add support for select_related and prefetch_related to GraphQL
The GraphQL queries for the TestGroup UI were creating too many sub-queries.
There is a bug in graphene_django that eliminates any usage of
select_related or prefetch_related.  This works around that limitation and
allows the user to create a mapping of fields to the optimization that
can be used for each.  It will only apply the optimization if the field
is present in the query.