Bug 1215587 - Add job_group field to jobs table (#2601)
This is to allow any job type to be able to belong to any
job group.  This will also mean that if someone accidentally
picked the wrong group for a job type, we don't need to
fix it in the DB for all new jobs.  They can fix their task
definition, and all new jobs will go to the new job group.

This includes a management command to migrate the old
data from job_type.job_group to the new field of job.job_group.

A follow-up PR will remove the old field and set the API to
read from the job.job_group field.