bug 1393277 - use --option=arg rather than --option arg
We currently use `--option=arg` in the decision task [1], but
`--option arg` in treeherder actions. These are essentially equivalent.
In scriptworkers' chain of trust verification, we verify the
action or decision task's command is allowlisted. To do so, we ignore
options [2]. By specifying them as `--option=arg`, we're able to skip
verifying args as well as options.

We have plans to make this better, but until then, let's use
`--option=arg`.

[1] https://hg.mozilla.org/mozilla-central/file/ab2d700fda2b/.taskcluster.yml#l111
[2] https://github.com/mozilla-releng/scriptworker/blob/8437d4cd38abe30bcfd4f8ff4dc40615481fd490/scriptworker/cot/verify.py#L903